requirements.txt
GitPython==3.1.31
inquirer==3.1.3
langchain==0.0.180
nbformat==5.8.0
numpy==1.23.5
python-dotenv==1.0.0
Requests==2.31.0